Added saving options to tango

+++ cryocat/app/layout/motlio.py
@@ -12,9 +12,25 @@ import dash_bootstrap_components as dbc
 from cryocat.cryomotl import Motl
 from cryocat.classutils import get_class_names_by_parent
 from cryocat.app.globalvars import tomo_ids
-from cryocat.app.apputils import get_print_out, save_output
+from cryocat.app.apputils import get_print_out, save_output, save_motl
+from cryocat.app.layout.customel import InlineLabeledDropdown, InlineInputForm
+
 
 # motl_types = [{"label": name, "value": name} for name in get_class_names_by_parent("Motl", "cryocat.cryomotl")]
+
+motl_types = [
+    {"label": "EM", "value": "emmotl"},
+    {"label": "STOPGAP", "value": "stopgap"},
+    {"label": "Relion", "value": "relion"},
+    {"label": "Dynamo", "value": "dynamo"},
+]
+
+relion_versions = [
+    {"label": "Version 3.0", "value": 3.0},
+    {"label": "Version 3.1", "value": 3.1},
+    {"label": "Version 4.x", "value": 4.0},
+    {"label": "Version 5.x", "value": 5.0},
+]
 
 
 def get_motl_save_component(prefix: str):
@@ -38,43 +54,25 @@ def get_motl_save_component(prefix: str):
                         [
                             html.Div(
                                 [
-                                    dcc.Dropdown(
-                                        id=f"{prefix}-datasave-dropdown",
+                                    InlineLabeledDropdown(
+                                        id_=f"{prefix}-datasave-dropdown",
+                                        label="Data to save:",
                                         multi=False,
                                         placeholder="Select data to save",
-                                        style={
-                                            "width": "100%",
-                                            # "height": "20px",  # reduce height
-                                            # "fontSize": "1.3rem",  # smaller font
-                                            "padding": "0",  # reduce padding
-                                            "marginBottom": "0.5rem",
-                                        },
                                     ),
-                                    dcc.Dropdown(
-                                        id=f"{prefix}-assignment-dropdown",
+                                    InlineLabeledDropdown(
+                                        id_=f"{prefix}-assignment-dropdown",
+                                        label="Store in the column:",
                                         multi=False,
                                         options=Motl.motl_columns,
                                         placeholder="Select column to store the output",
-                                        style={
-                                            "width": "100%",
-                                            # "height": "20px",  # reduce height
-                                            # "fontSize": "1.3rem",  # smaller font
-                                            "padding": "0",  # reduce padding
-                                            "marginBottom": "0.5rem",
-                                        },
                                     ),
-                                    dcc.Dropdown(
-                                        id=f"{prefix}-data-save-dropdown",
+                                    InlineLabeledDropdown(
+                                        id_=f"{prefix}-data-save-dropdown",
                                         options=["All", "Specific classes"],
+                                        label="Select class:",
                                         multi=False,
-                                        placeholder="What to save",
-                                        style={
-                                            "width": "100%",
-                                            # "height": "20px",  # reduce height
-                                            # "fontSize": "1.3rem",  # smaller font
-                                            "padding": "0",  # reduce padding
-                                            "marginBottom": "0.5rem",
-                                        },
+                                        placeholder="Select what to save",
                                     ),
                                     dcc.Checklist(
                                         options=[],
@@ -91,8 +89,44 @@ def get_motl_save_component(prefix: str):
                                             "marginBottom": "0.5rem",
                                         },
                                     ),
-                                    dbc.Input(
-                                        id=f"{prefix}-save-path-input",
+                                    InlineLabeledDropdown(
+                                        id_=f"{prefix}-data-save-type-dropdown",
+                                        options=motl_types,
+                                        label="Output type:",
+                                        multi=False,
+                                        placeholder="Output type",
+                                    ),
+                                    InlineLabeledDropdown(
+                                        id_=f"{prefix}-save-motl-relion-version-dropdown",
+                                        label="Version:",
+                                        default_visibility="hidden",
+                                        options=relion_versions,
+                                    ),
+                                    html.Div(
+                                        id=f"{prefix}-save-motl-relion-options",
+                                        className="hidden",
+                                        children=[
+                                            InlineInputForm(
+                                                id_=f"{prefix}-save-motl-relion-pixelsize",
+                                                type="number",
+                                                placeholder="Pixel size",
+                                                min=1.0,
+                                                step=1,
+                                                label="Pixel size (A):",
+                                            ),
+                                            InlineInputForm(
+                                                id_=f"{prefix}-save-motl-relion-binning",
+                                                type="number",
+                                                placeholder="Binning",
+                                                min=1.0,
+                                                step=1,
+                                                label="Binning:",
+                                            ),
+                                        ],
+                                    ),
+                                    InlineInputForm(
+                                        id_=f"{prefix}-save-path-input",
+                                        label="Filename:",
                                         type="text",
                                         placeholder="Filename (including its path)",
                                     ),
@@ -172,6 +206,34 @@ def register_motl_save_callbacks(prefix: str, stored_outputs, connected_store_id
         return options
 
     @callback(
+        Output(f"{prefix}-save-motl-relion-version-dropdown", "value", allow_duplicate=True),
+        Output(f"{prefix}-save-motl-relion-version-dropdown-topdiv", "className", allow_duplicate=True),
+        Output(f"{prefix}-save-motl-relion-options", "className", allow_duplicate=True),
+        Input(f"{prefix}-data-save-type-dropdown", "value"),
+        prevent_initial_call=True,
+    )
+    def display_options_motl_type(motl_type):
+
+        if motl_type == "relion":
+            return "Version 3.0", "flex", "hidden"
+        else:
+            return "Version 3.0", "hidden", "hidden"
+
+    @callback(
+        Output(f"{prefix}-save-motl-relion-options", "className", allow_duplicate=True),
+        Input(f"{prefix}-save-motl-relion-version-dropdown", "value"),
+        prevent_initial_call=True,
+    )
+    def display_options_relion_version(relion_version):
+
+        if relion_version == "Version 3.0" or relion_version == "Version 3.1":
+            return "hidden"
+        elif relion_version == "Version 4.x":
+            return "flex"
+        else:
+            return "flex"
+
+    @callback(
         Output(f"{prefix}-status-label", "children", allow_duplicate=True),
         Input(f"{prefix}-save-output-file", "n_clicks"),
         State(f"{prefix}-datasave-dropdown", "value"),
@@ -180,11 +242,32 @@ def register_motl_save_callbacks(prefix: str, stored_outputs, connected_store_id
         State(f"{prefix}-classes-checklist", "value"),
         State(f"{prefix}-classes-checklist", "options"),
         State(connected_store_id, "data"),
+        State(f"{prefix}-data-save-type-dropdown", "value"),
+        State(f"{prefix}-motl-extra-data-store", "data"),
+        State(f"{prefix}-relion-optics-store", "data"),
+        State(f"{prefix}-relion5-tomos-store", "data"),
+        State(f"{prefix}-save-motl-relion-binning", "value"),
+        State(f"{prefix}-save-motl-relion-pixelsize", "value"),
+        State(f"{prefix}-save-motl-relion-version-dropdown", "value"),
         store_states,
         prevent_initial_call=True,
     )
     def save_data(
-        n_clicks, data_type_value, file_path, column_id, class_filter, checklist_options, data_to_save, *store_states
+        n_clicks,
+        data_type_value,
+        file_path,
+        column_id,
+        class_filter,
+        checklist_options,
+        data_to_save,
+        motl_type,
+        extra_df_data,
+        relion_optics,
+        relion_tomos,
+        relion_binning,
+        relion_pixel_size,
+        relion_version,
+        *store_states,
     ):
 
         data_index = list(stored_outputs.keys()).index(data_type_value)
@@ -226,7 +309,17 @@ def register_motl_save_callbacks(prefix: str, stored_outputs, connected_store_id
         # Drop rows with no match (i.e., column_id is NaN)
         motl_df = motl_df.dropna(subset=[column_id])
 
-        status = save_output(file_path=file_path, data_to_save=motl_df, csv_only=False)
+        status = save_motl(
+            file_path=file_path,
+            data_to_save=motl_df,
+            motl_type=motl_type,
+            extra_df=extra_df_data,
+            rln_optics=relion_optics,
+            rln_tomos=relion_tomos,
+            rln_binning=relion_binning,
+            rln_pixel_size=relion_pixel_size,
+            rln_version=relion_version,
+        )
 
         return status
 
@@ -250,9 +343,9 @@ def get_motl_load_component(prefix: str, display_option="block"):
                     dbc.Col(
                         dcc.Dropdown(
                             id=f"{prefix}-motl-dropdown",
-                            options=["EM motl", "STOPGAP", "Relion", "Dynamo"],
+                            options=motl_types,
                             multi=False,
-                            value="EM motl",
+                            value="emmotl",
                             style={
                                 "width": "100%",
                                 # "height": "20px",  # reduce height
@@ -269,7 +362,8 @@ def get_motl_load_component(prefix: str, display_option="block"):
                     id=f"{prefix}-motl-relion-version-dropdown",
                     options=["Version 3.0", "Version 3.1", "Version 4.x", "Version 5.x"],
                 ),
-                style={"display": "none"},
+                style={"marginTop": "1rem"},
+                className="hidden",
                 id=f"{prefix}-motl-relion-version",
             ),
             dbc.Row(
@@ -321,7 +415,8 @@ def get_motl_load_component(prefix: str, display_option="block"):
                         id=f"{prefix}-motl-relion-tomos-display",
                     ),
                 ],
-                style={"display": "none"},
+                style={"marginTop": "1rem"},
+                className="hidden",
                 id=f"{prefix}-motl-relion-options",
             ),
             dbc.Row(
@@ -361,6 +456,9 @@ def register_motl_load_callbacks(prefix: str):
 
     @callback(
         Output(f"{prefix}-motl-data-store", "data", allow_duplicate=True),
+        Output(f"{prefix}-motl-extra-data-store", "data", allow_duplicate=True),
+        Output(f"{prefix}-relion-optics-store", "data", allow_duplicate=True),
+        Output(f"{prefix}-relion5-tomos-store", "data", allow_duplicate=True),
         Input(f"{prefix}-motl-upload", "contents"),
         State(f"{prefix}-motl-upload", "filename"),
         State(f"{prefix}-motl-dropdown", "value"),
@@ -380,14 +478,23 @@ def register_motl_load_callbacks(prefix: str):
             tmp_file_path = tmp_file.name
             tmp_file.close()
 
-        motl_type = motl_type.lower().replace(" ", "")
+        extra_data = None
+        relion_optics = None
+        relion_tomos = None
+
         if motl_type != "relion":
+
             motl = Motl.load(tmp_file_path, motl_type)
+
+            if motl_type == "stopgap":
+                extra_data = motl.stopgap_df.to_dict("records")
+            elif motl_type == "dynamo":
+                extra_data = motl.dynamo_df.to_dict("records")
         else:
             if rln_version == "Version 3.0":
                 rln_kwargs = {"version": 3.0}
             elif rln_version == "Version 3.1":
-                rln_kwargs = {"version": 3.0}
+                rln_kwargs = {"version": 3.1}
             elif rln_version == "Version 4.x":
                 rln_kwargs = {"version": 4.0, "pixel_size": rln_pixelsize, "binning": rln_binning}
             else:
@@ -405,8 +512,13 @@ def register_motl_load_callbacks(prefix: str):
                         rln_kwargs["input_tomograms"] = tomo_tmp_file_path
 
             motl = Motl.load(tmp_file_path, motl_type, **rln_kwargs)
-            if rln_tomos:
-                os.remove(tomo_tmp_file_path)
+            extra_data = motl.relion_df.to_dict("records")
+            relion_optics = motl.optics_data.to_dict("records")
+
+            if motl_type == "relion5":
+                relion_tomos = motl.tomo_df.to_dict("records")
+                if rln_tomos:
+                    os.remove(tomo_tmp_file_path)
 
         os.remove(tmp_file_path)
 
@@ -416,33 +528,33 @@ def register_motl_load_callbacks(prefix: str):
         file_status = f"Loaded {filename};   " + get_print_out(motl)
         table_data = motl.df.to_dict("records")
 
-        return table_data
+        return table_data, extra_data, relion_optics, relion_tomos
 
     @callback(
         Output(f"{prefix}-motl-relion-version-dropdown", "value", allow_duplicate=True),
-        Output(f"{prefix}-motl-relion-version", "style", allow_duplicate=True),
-        Output(f"{prefix}-motl-relion-options", "style", allow_duplicate=True),
+        Output(f"{prefix}-motl-relion-version", "className", allow_duplicate=True),
+        Output(f"{prefix}-motl-relion-options", "className", allow_duplicate=True),
         Input(f"{prefix}-motl-dropdown", "value"),
         prevent_initial_call=True,
     )
     def display_options_motl_type(motl_type):
 
-        if motl_type == "Relion":
-            return "Version 3.0", {"display": "flex", "marginTop": "1rem"}, {"display": "none"}
+        if motl_type == "relion":
+            return "Version 3.0", "flex", "hidden"
         else:
-            return "Version 3.0", {"display": "none"}, {"display": "none"}
+            return "Version 3.0", "hidden", "hidden"
 
     @callback(
-        Output(f"{prefix}-motl-relion-options", "style", allow_duplicate=True),
-        Output(f"{prefix}-motl-relion-tomos-display", "style", allow_duplicate=True),
+        Output(f"{prefix}-motl-relion-options", "className", allow_duplicate=True),
+        Output(f"{prefix}-motl-relion-tomos-display", "className", allow_duplicate=True),
         Input(f"{prefix}-motl-relion-version-dropdown", "value"),
         prevent_initial_call=True,
     )
     def display_options_relion_version(relion_version):
 
         if relion_version == "Version 3.0" or relion_version == "Version 3.1":
-            return {"display": "none"}, {"display": "none"}
+            return "hidden", "hidden"
         elif relion_version == "Version 4.x":
-            return {"display": "flex", "marginTop": "1rem"}, {"display": "none"}
+            return "flex", "hidden"
         else:
-            return {"display": "flex", "marginTop": "1rem"}, {"display": "flex", "marginTop": "1rem"}
+            return "flex", "flex"
